+++ ShakerSort.hpp
@@ -2,9 +2,9 @@
 #include <chrono>
 #include <vector>
 
-std::pair<long long, double> ShakerSort(int a[], int n)
+std::pair<double, double> ShakerSort(int a[], int n)
 {
-    int compare = 0;
+    double compare = 0;
     auto start = std::chrono::high_resolution_clock::now();
     int Left = 0;
     int Right = n - 1;
